Update Main.java
Creates rudimentary dialogue that prompts the user with four options: "fluff, disco, fluffy disco," or "neither." 

If the input matches either of the first three options, regardless of capitalization, the program responds with the appropriate comment. Otherwise, the program responds with a default option.

Once the program has completed the small interaction, it closes the scanner and ends the program.